20250526/python/pyproject.toml
[project]
name = "slack-bot-demo"
version = "0.1.0"
description = "Slack Bot Demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.16.1",
    "python-dotenv>=1.1.0",
    "slack-bolt>=1.23.0",
    "slack-sdk>=3.35.0",
]

[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["I"]
